refactor: added index.js to data folder
