+++ src/data/index.js
@@ -1,0 +1,6 @@
+import { brandLogoData } from "./brandCategoryCover.data";
+import { categorySectionItems } from "./categoryCover.data";
+import { interestCoverData } from "./interestCategoryCover.data";
+import { newArrivalCover } from "./newCategoryCover.data";
+
+export {brandLogoData, categorySectionItems, interestCoverData, newArrivalCover}
