+++ src/pages/Home/components/CategorySection.jsx
@@ -1,4 +1,4 @@
-import { categorySectionItems } from "../../../data/categoryCover.data"
+import { categorySectionItems } from "../../../data"
 
 export const CategorySection = () => {
